+++ cleaner/src/cleaner.rs
@@ -65,6 +65,10 @@ pub fn clean_doc(
     options: &CleaningOptions,
     opt: &WriteOptions
 ) -> Result<(), error::Error> {
+    if options.remove_nonsvg_attributes {
+        remove_nonsvg_attributes(doc);
+    }
+
     preclean_checks(doc)?;
 
     // NOTE: Order is important.